refactor(ruby): extract Rails/route DSL edge emitters into rails.go
Pure move: the include/association/callback/scope/broadcast/importmap
emitters and the whole route-helper family leave the walk core for their
own concern file. ruby.go is now a ~310-line walk-core (Extract, walk,
dispatchCall, block helpers). Behavior-preserving: 24 goldens byte-identical.
